Refactor query engine.
This refactor moves the query generation to a new module and allows
users to define queries before using them. It also moves the retry logic
into a general location. Actions are now located in `Wallaby.Browser`.
The footprint of available actions is diminished because we no longer
need actions for every type of query.